refactor: user service, repository layer name
